improved grasp detection for the ArUco cube
This commit improves grasp detection for the standard ArUco cube. Sometimes the robot would successfully grasp the ArUco cube, but not detect that the grasp was successful. As a result, the gripper would not be retracted with the held ArUco cube.

The grasp detector uses the following criteria:
 1. The distance between the point between the fingertips and the center of the cube is small.
 2. The gripper's actuator effort has a sufficiently large magnitude with the correct sign.
 3. The distance between the fingertips is larger than a minimum value.
 4. The distance between the fingertips is smaller than a maximum value.

Testing indicated that criterion 4 was resulting in failed grasp detection. In other words, the estimated distance between the fingertips was too large to be considered a successful grasp.

This commit increases the allowable distance between the fingertips during a successful grasp. The prior maximum value was 7.5cm. Now the maximum value is 8.5cm.

During a brief test, this 1cm increase resulted in all ArUco cube grasps being successfully detected and did not result in false positives.

During a brief test, this change did not appear to significantly affect performance with a tennis ball.