recenter with position control
When started, the robot now uses position control to recenter itself prior to visually servoing. It waits until the robot has achieved the center pose to begin visually servoing.

Previously, the robot used velocity control to recenter itself. This was not a reliable process. Issues included lack of precision for the final joint poses and sensitivity to termination conditions. In addition, there appeared to be timing and parallelism issues likely due to starting up the robot, sensors, and networking. Sometimes, the robot would fail to stop the wrist yaw, resulting in the wrist yaw hitting a hard stop.