lower error starts grasp & use median depth
This commit makes two changes in an attempt to address issues that have been experienced on other robots. At the time of making this commit, these two changes have only been tested on a single Stretch 3.

Grasping will not begin until the distance between the estimated center of the fingertips and the estimated center of the object is less than 2 cm. Previously, the robot would begin grasping with an estimated error of less than 4.5 cm. This change is intended to increase grasping precision.

When YOLOv8 segments a tennis ball, the median depth for the pixels in the segmented mask is used. Previously, the mean depth was used. The median depth should be less influenced by outliers.